updated buildGraph
buildGraph now properly builds the graph based off of the CSV file. When App.java is run, put in directory of the file ( i.e. C:\Users\JC\Desktop\CS-310\Program3\layout.csv ) and it will add the vertices and make the connections. For now, it prints to the console the line of the current BufferedReader and whether it is adding a new vertex or making a connection.